createQRcode_test/requirements.txt
qrcode==7.4.2
pillow==10.3.0
Flask
Gunicorn
requests
